Try doing Offscreen with Multiple Scenes
I forgot where I was but if I recall correctly
you can't use the method where put the canvas in back
and draw the objects to match the scene because it will
always lag.

So, instead you need to use an OffscreenCanvas and
transferToBitmap but I'm 99% sure that's a bad method.
The problem is say you have 100 objects, you'll have
100 bitmap canvases. You can save work by only transfering
those that are on the screen but you still have others
sitting around. I suppose you could keep a cache of bitmap
canvases and shuffle them around in the main thread, telling
the worker which ones to update. The problem there is
when you moved a canvas you'd get the wrong image in it for
a frame. I suppose you could attach a clear bitmap to it

Anyway, ...